requirements.txt
youtube-transcript-api>=0.6.0
requests>=2.28.0
google-genai>=1.0.0
python-dotenv>=1.0.0
